Refactored RightSidebar into functions, context
